fix: rename passenger-dahsboard.service.ts to pokedex-dashboard.service.ts
